Fix upvote duplication bug
We used to check for already submitted links of identities by using
lmdbx-js db.doesExist(key). However, in seldom cases, it seemed to have
matched shorter links greedily such that it wasn't possible to upvote
certain links anymore.

There were additional problems with the constraints database that we
fixed in this update

- the terminal symbol was the colon, which is akward considering that a
  url scheme also contains one. We're replacing the colon with a pipe.
- We found some constraints that didn't have any corresponding posts, so
  we remove them.
- We're now normalizing the URL