Remove @ledgerhq/connect-kit-loader for good
As we had been using rainbowkit's getDefaultWallets, from what it
currently looks like Kiwi News was never exposed to the vulnerability
because the Ledger connecter was simply never invoked in the code.
WalletConnect seems to not have been affected by the vulnerability.

With this fix, we've now shrinkwrapped the package-lock file and removed
the connect-kit-loader package. The connect buttons etc. are re-enabled.

- https://twitter.com/wevm_dev/status/1735290192599110047
- https://twitter.com/WalletConnect/status/1735299408130933059