Transform metadataDB into a memory cache
We initially had designed the metadata database as an lmdb instance
running side to side with the ethereumjs-trie. However, as this commit
also shows, this ended up putting in question the authority of the trie
as the source of truth with the metadata database sometimes making the
call on whether an upvote was valid to go into the database.

The benefits of using a database for this was that the constraints data
of who had upvoted was immediately upon startup available and that this
concept is generally scalable for a large amount of data.

However, one really painful property of storing this on disk and having
to keep the metadata database synchronized with the trie was that we had
several faults where either the trie or the metadata database was
rolling back some uncommitted state which lead to PERMANENT
inconsistencies between the databases. We initially addressed this by
introducing naive rollback logic using the try catch pattern, but as is
also visible in one of the now-deleted comments, try catch cannot really
address these issues when the node program panicing is a realistically
occurring fault.

Hence, removing the necessity to write to disk asynchronously simplifies
the storage operation significantly. By keeping the data in memory, this
also means that potential inconsistently through faults are merely
temporary until the next node restart - which at the moment is anyways
every few days because of rolling updates.