Improve local account selection algorithm
In some cases, a user would set up an application key but then transfer
the NFT through their custody wallet. In this case, the application key
would still be present, but it became in-eligible already. Hence, we're
now looking at the respective identity of that key before setting it as
the application's preferred signer.