Improve safety of cache entry
It seems a comment can have a lower timestamp than the submission? This
is strange because it would point to inconsistencies in the storage
algorithm.

In any case, considering this could also be a local artifact, it is
important to just have that part of the re-creation algorithm of the
cache fail while the rest continues to run.